move ReviewDashboardSkills out of app-review
