+++ packages/@coorpacademy-components/src/template/review-dashboard-skills/style.css
