+++ packages/@coorpacademy-components/src/template/review-dashboard-skills/test/fixtures/loading.js
@@ -1,0 +1,10 @@
+import reviewPresentation from '../../../../atom/review-presentation/test/fixtures/default';
+import reviewSkills from '../../../../organism/review-skills/test/fixtures/loading';
+
+export default {
+  props: {
+    'aria-label': 'Review Skills Template',
+    reviewPresentation: reviewPresentation.props,
+    reviewSkills: reviewSkills.props
+  }
+};
